chore(sdk-coin-near): move account-lib tests
- moves unit tests & resources from account-lib

Ticket: BG-48154
