+++ modules/sdk-coin-near/test/resources/near.ts
