+++ modules/sdk-coin-near/test/unit/getBuilderFactory.ts
@@ -1,0 +1,6 @@
+import { TransactionBuilderFactory } from '../../src';
+import { coins } from '@bitgo/statics';
+
+export const getBuilderFactory = (coin: string): TransactionBuilderFactory => {
+  return new TransactionBuilderFactory(coins.get(coin));
+};
